refactor: port PR67 story service audit fixes
