+++ LESSONS_LEARNED.md
@@ -61,4 +61,6 @@ This file consolidates lessons that should shape the PR #70 recovery and future 
 - Agent instructions must be corrected before long-running recovery work. Otherwise future agents will follow stale deployment assumptions and undo the user's Vercel direction.
 - Angular 20 in this checkout should be verified with Node 20. Local Node v23.8.0 can abort during `ng build` without useful diagnostics, while the same branch builds under temporary Node v20.20.2.
 - PR #70 showed a concrete Angular-template lesson: keep object spread, arrow functions, and complex state updates out of templates. Put that logic on the component class so Angular's template compiler and type checker can reason about it.
-- AI generation configuration needs executable verification. `tests/verify-ai-fixes.test.ts` caught stale path assumptions and now checks model name, token budgeting, API parameters, and timeout shape across both story service copies.
+- AI generation configuration needs executable verification. `tests/verify-ai-fixes.test.ts` caught stale path assumptions and now checks model name, token budgeting, API parameters, and timeout shape against the canonical story service.
+- A passing process exit is not enough evidence when a custom test harness is involved. The #67 pass found a test that printed a failure but exited 0 because it counted a non-existent `failed` property.
+- Duplicate service deletion should happen before deeper feature ports. Keeping `story-generator/src/api/lib` alive made every story-generation fix twice as risky and encouraged stale-path imports.
