--- story-generator/src/api/lib/types/contracts.ts
@@ -1,391 +1,0 @@
-// ==================== SEAM-DRIVEN DEVELOPMENT CONTRACTS ====================
-// These contracts are derived directly from UI interactions and data flows
-// Each seam represents a boundary where data crosses between components
-
-// ==================== TYPE DEFINITIONS ====================
-export type CreatureType = 'vampire' | 'werewolf' | 'fairy';
-export type ThemeType = 'betrayal' | 'obsession' | 'power_dynamics' | 'forbidden_love' | 'revenge' | 'manipulation' | 'seduction' | 'dark_secrets' | 'corruption' | 'dominance' | 'submission' | 'jealousy' | 'temptation' | 'sin' | 'desire' | 'passion' | 'lust' | 'deceit';
-export type SpicyLevel = 1 | 2 | 3 | 4 | 5;
-export type WordCount = 700 | 900 | 1200;
-export type VoiceType = 'female' | 'male' | 'neutral';
-export type CharacterVoiceType = 
-  | 'vampire_male' | 'vampire_female' 
-  | 'werewolf_male' | 'werewolf_female'
-  | 'fairy_male' | 'fairy_female'
-  | 'human_male' | 'human_female'
-  | 'narrator';
-export type AudioSpeed = 0.5 | 0.75 | 1.0 | 1.25 | 1.5;
-export type AudioFormat = 'mp3' | 'wav' | 'aac';
-export type ExportFormat = 'pdf' | 'txt' | 'html' | 'epub' | 'docx';
-export type ImageStyle = 'artistic' | 'photorealistic' | 'fantasy' | 'dark' | 'romantic';
-
-export interface AudioProgress {
-  percentage: number; // 0-100
-  status: 'queued' | 'processing' | 'completed' | 'failed';
-  message: string;
-  estimatedTimeRemaining?: number; // in seconds
-}
-
-// ==================== SEAM 1: USER INPUT → STORY GENERATOR ====================
-export interface StoryGenerationSeam {
-  seamName: "User Input → Story Generator";
-  description: "Converts form data into generated story content";
-
-  input: {
-    creature: CreatureType;
-    themes: ThemeType[];
-    userInput: string; // Optional custom ideas
-    spicyLevel: SpicyLevel;
-    wordCount: WordCount;
-  };
-
-  output: {
-    storyId: string;
-    title: string;
-    content: string; // HTML formatted content for [innerHTML] binding (speaker tags removed)
-    rawContent?: string; // Content with speaker tags for audio processing
-    creature: CreatureType;
-    themes: ThemeType[];
-    spicyLevel: SpicyLevel;
-    actualWordCount: number;
-    estimatedReadTime: number; // in minutes
-    hasCliffhanger: boolean; // determines if "Continue Chapter" button shows
-    generatedAt: Date;
-  };
-
-  errors: {
-    GENERATION_FAILED: {
-      code: "GENERATION_FAILED";
-      message: string;
-      retryable: boolean;
-      retryAfter?: number;
-    };
-    CONTENT_VIOLATION: {
-      code: "CONTENT_VIOLATION";
-      message: string;
-      suggestions: string[];
-      blockedContent: string[];
-    };
-    RATE_LIMITED: {
-      code: "RATE_LIMITED";
-      message: string;
-      retryAfter: number; // seconds
-      limitRemaining: number;
-    };
-    INVALID_INPUT: {
-      code: "INVALID_INPUT";
-      message: string;
-      field: keyof StoryGenerationSeam['input'];
-      providedValue: any;
-      expectedType: string;
-    };
-  };
-}
-
-// ==================== SEAM 2: STORY → CHAPTER CONTINUATION ====================
-export interface ChapterContinuationSeam {
-  seamName: "Story → Chapter Continuation";
-  description: "Extends existing story with additional chapters";
-
-  input: {
-    storyId: string;
-    currentChapterCount: number;
-    existingContent: string; // Full story HTML content
-    userInput?: string; // Optional continuation hints
-    maintainTone: boolean; // Keep same spicy level and themes
-  };
-
-  output: {
-    chapterId: string;
-    chapterNumber: number;
-    title: string;
-    content: string; // New chapter HTML content
-    wordCount: number;
-    cliffhangerEnding: boolean;
-    themesContinued: ThemeType[];
-    spicyLevelMaintained: SpicyLevel;
-    appendedToStory: string; // Full updated story content
-  };
-
-  errors: {
-    STORY_NOT_FOUND: {
-      code: "STORY_NOT_FOUND";
-      message: string;
-      storyId: string;
-    };
-    CONTINUATION_FAILED: {
-      code: "CONTINUATION_FAILED";
-      message: string;
-      retryable: boolean;
-    };
-    MAX_CHAPTERS_REACHED: {
-      code: "MAX_CHAPTERS_REACHED";
-      message: string;
-      maxChapters: number;
-      currentChapters: number;
-    };
-  };
-}
-
-// ==================== SEAM 2.5: REAL-TIME STORY GENERATION ====================
-export interface StreamingStoryGenerationSeam {
-  seamName: "Real-Time Story Generation";
-  description: "Provides real-time updates during story generation for better UX";
-
-  input: {
-    // Same as StoryGenerationSeam input
-    creature: CreatureType;
-    themes: ThemeType[];
-    userInput: string;
-    spicyLevel: SpicyLevel;
-    wordCount: WordCount;
-  };
-
-  progressUpdate: {
-    streamId: string;
-    storyId?: string; // Generated after stream starts
-    type: 'connected' | 'progress' | 'chunk' | 'complete' | 'error';
-    content?: string; // Accumulated content so far
-    isComplete: boolean;
-    metadata: {
-      wordsGenerated: number;
-      totalWordsTarget: number;
-      estimatedWordsRemaining: number;
-      generationSpeed: number; // words per second
-      percentage: number; // 0-100
-      estimatedTimeRemaining?: number; // seconds
-    };
-  };
-
-  finalOutput: {
-    // Same as StoryGenerationSeam output
-    storyId: string;
-    title: string;
-    content: string;
-    rawContent?: string;
-    creature: CreatureType;
-    themes: ThemeType[];
-    spicyLevel: SpicyLevel;
-    actualWordCount: number;
-    estimatedReadTime: number;
-    hasCliffhanger: boolean;
-    generatedAt: Date;
-  };
-
-  errors: {
-    STREAMING_CONNECTION_FAILED: {
-      code: "STREAMING_CONNECTION_FAILED";
-      message: string;
-      retryable: boolean;
-    };
-    STREAM_INTERRUPTED: {
-      code: "STREAM_INTERRUPTED";
-      message: string;
-      recoveryOptions: string[];
-      partialContent?: string;
-    };
-    // Inherits all errors from StoryGenerationSeam
-  };
-}
-
-// ==================== SEAM 3: STORY TEXT → AUDIO CONVERTER ====================
-export interface AudioConversionSeam {
-  seamName: "Story Text → Audio Converter";
-  description: "Converts story text to audio format with progress tracking";
-
-  input: {
-    storyId: string;
-    content: string; // Full story HTML content
-    voice?: VoiceType;
-    speed?: AudioSpeed;
-    format?: AudioFormat;
-  };
-
-  output: {
-    audioId: string;
-    storyId: string;
-    audioUrl: string; // Download URL for UI
-    duration: number; // in seconds
-    fileSize: number; // in bytes
-    format: AudioFormat;
-    voice: VoiceType;
-    speed: AudioSpeed;
-    progress: AudioProgress; // For real-time UI updates
-    completedAt: Date;
-  };
-
-  errors: {
-    CONVERSION_FAILED: {
-      code: "CONVERSION_FAILED";
-      message: string;
-      retryable: boolean;
-      stage: "text_processing" | "audio_generation" | "file_creation";
-    };
-    UNSUPPORTED_CONTENT: {
-      code: "UNSUPPORTED_CONTENT";
-      message: string;
-      unsupportedElements: string[];
-    };
-    AUDIO_QUOTA_EXCEEDED: {
-      code: "AUDIO_QUOTA_EXCEEDED";
-      message: string;
-      quotaRemaining: number;
-      resetTime: Date;
-    };
-  };
-}
-
-// ==================== SEAM 4: STORY DATA → SAVE/EXPORT SYSTEM ====================
-export interface SaveExportSeam {
-  seamName: "Story Data → Save/Export System";
-  description: "Saves or exports story data in various formats";
-
-  input: {
-    storyId: string;
-    content: string; // Full story HTML content
-    title: string;
-    format: ExportFormat;
-    includeMetadata?: boolean;
-    includeChapters?: boolean;
-  };
-
-  output: {
-    exportId: string;
-    storyId: string;
-    downloadUrl: string; // Direct download URL for UI
-    filename: string;
-    format: ExportFormat;
-    fileSize: number;
-    expiresAt: Date; // When download URL expires
-    exportedAt: Date;
-  };
-
-  errors: {
-    EXPORT_FAILED: {
-      code: "EXPORT_FAILED";
-      message: string;
-      retryable: boolean;
-      format: ExportFormat;
-    };
-    FORMAT_NOT_SUPPORTED: {
-      code: "FORMAT_NOT_SUPPORTED";
-      message: string;
-      requestedFormat: ExportFormat;
-      supportedFormats: ExportFormat[];
-    };
-    STORAGE_QUOTA_EXCEEDED: {
-      code: "STORAGE_QUOTA_EXCEEDED";
-      message: string;
-      quotaRemaining: number;
-    };
-  };
-}
-
-// ==================== SEAM 5: STORY → IMAGE GENERATION ====================
-export interface ImageGenerationSeam {
-  seamName: "Story → Image Generation";
-  description: "Generates images based on story content using Grok-2-Image";
-
-  input: {
-    storyId: string;
-    content: string; // Story content or specific scene
-    imagePrompt?: string; // Optional custom prompt
-    creature: CreatureType;
-    themes: ThemeType[];
-    style: ImageStyle;
-    aspectRatio?: '1:1' | '16:9' | '9:16' | '4:3';
-  };
-
-  output: {
-    imageId: string;
-    storyId: string;
-    imageUrl: string; // URL to generated image
-    prompt: string; // The actual prompt sent to AI
-    style: ImageStyle;
-    aspectRatio: string;
-    width: number;
-    height: number;
-    fileSize: number; // in bytes
-    generatedAt: Date;
-  };
-
-  errors: {
-    IMAGE_GENERATION_FAILED: {
-      code: "IMAGE_GENERATION_FAILED";
-      message: string;
-      retryable: boolean;
-      reason: "content_policy" | "quota_exceeded" | "service_error";
-    };
-    UNSUPPORTED_STYLE: {
-      code: "UNSUPPORTED_STYLE";
-      message: string;
-      requestedStyle: ImageStyle;
-      supportedStyles: ImageStyle[];
-    };
-    IMAGE_QUOTA_EXCEEDED: {
-      code: "IMAGE_QUOTA_EXCEEDED";
-      message: string;
-      quotaRemaining: number;
-      resetTime: Date;
-    };
-  };
-}
-
-// ==================== VALIDATION RULES ====================
-export const VALIDATION_RULES = {
-  userInput: {
-    maxLength: 1000,
-    allowedHtml: false
-  },
-  themes: {
-    maxCount: 5,
-    allowedValues: ['betrayal', 'obsession', 'power_dynamics', 'forbidden_love', 'revenge', 'manipulation', 'seduction', 'dark_secrets', 'corruption', 'dominance', 'submission', 'jealousy', 'temptation', 'sin', 'desire', 'passion', 'lust', 'deceit']
-  },
-  spicyLevel: {
-    min: 1,
-    max: 5
-  },
-  wordCount: {
-    allowedValues: [700, 900, 1200]
-  },
-  audioSpeed: {
-    min: 0.5,
-    max: 1.5
-  },
-  imageStyle: {
-    allowedValues: ['artistic', 'photorealistic', 'fantasy', 'dark', 'romantic']
-  },
-  aspectRatio: {
-    allowedValues: ['1:1', '16:9', '9:16', '4:3']
-  }
-} as const;
-
-// ==================== UI STATE MANAGEMENT ====================
-export interface UIState {
-  isGenerating: boolean;
-  isConvertingAudio: boolean;
-  isSaving: boolean;
-  isGeneratingNext: boolean;
-  isGeneratingImage: boolean;
-  audioProgress: number;
-  saveSuccess: boolean;
-  audioSuccess: boolean;
-  imageSuccess: boolean;
-  lastError?: string;
-}
-
-// ==================== UNIFIED API RESPONSE ====================
-export interface ApiResponse<T> {
-  success: boolean;
-  data?: T;
-  error?: {
-    code: string;
-    message: string;
-    details?: any;
-  };
-  metadata?: {
-    requestId: string;
-    processingTime: number;
-    rateLimitRemaining?: number;
-  };
-}
